requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
darkdetect==0.8.0
emoji-country-flag==1.3.2
idna==3.6
pillow==10.2.0
requests==2.31.0
urllib3==2.2.0
